requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
PyJWT==2.10.0
python-decouple==3.8
sqlparse==0.5.2
typing_extensions==4.12.2
